Extract class to test proposal cancel.
