+++ opengever/meeting/tests/test_proposal.py
@@ -644,22 +644,6 @@ class TestProposal(IntegrationTestCase):
                          proposal_model.submitted_title)
 
     @browsing
-    def test_proposal_can_be_cancelled(self, browser):
-        self.login(self.dossier_responsible, browser)
-        self.assertEqual(Proposal.STATE_PENDING, self.draft_proposal.get_state())
-        self.assert_workflow_state('proposal-state-active', self.draft_proposal)
-
-        browser.open(self.draft_proposal, view='tabbedview_view-overview')
-        browser.click_on("proposal-transition-cancel")
-        browser.click_on("Confirm")
-
-        statusmessages.assert_no_error_messages()
-        statusmessages.assert_message('Review state changed successfully.')
-
-        self.assertEqual(Proposal.STATE_CANCELLED, self.draft_proposal.get_state())
-        self.assert_workflow_state('proposal-state-cancelled', self.draft_proposal)
-
-    @browsing
     def test_proposal_can_be_reactivated(self, browser):
         self.login(self.dossier_responsible, browser)
         api.content.transition(self.draft_proposal, 'proposal-transition-cancel')
